cuda wip

+++ main.cpp
@@ -27,7 +27,7 @@ int main() {
     net.set_random();
 
     // define targets
-    std::vector<std::vector<uint8_t>> target = {
+    std::vector<std::vector<uint32_t>> target = {
         {1, 0, 0, 0, 0, 0, 0, 0, 0, 0}, {0, 1, 0, 0, 0, 0, 0, 0, 0, 0},
         {0, 0, 1, 0, 0, 0, 0, 0, 0, 0}, {0, 0, 0, 1, 0, 0, 0, 0, 0, 0},
         {0, 0, 0, 0, 1, 0, 0, 0, 0, 0}, {0, 0, 0, 0, 0, 1, 0, 0, 0, 0},
@@ -36,9 +36,9 @@ int main() {
 
     // get storage
     std::vector<std::vector<float>> test_images(N_TEST);
-    std::vector<std::vector<uint8_t>> test_labels(N_TEST);
+    std::vector<std::vector<uint32_t>> test_labels(N_TEST);
     std::vector<std::vector<float>> train_images(N_TRAIN);
-    std::vector<std::vector<uint8_t>> train_labels(N_TRAIN);
+    std::vector<std::vector<uint32_t>> train_labels(N_TRAIN);
 
     // load test data
     TIC(load_data)
